corrected the previous warning fix so that it doesn't break the code
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@14856 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/html/htmltag.cpp
@@ -78,9 +78,9 @@ wxHtmlTagsCache::wxHtmlTagsCache(const wxString& source)
             tg = m_CacheSize++;
             m_Cache[tg].Key = stpos = pos++;
 
-            size_t i;
+            int i;
             for ( i = 0;
-                  pos < lng && i < WXSIZEOF(tagBuffer) - 1 &&
+                  pos < lng && i < (int)WXSIZEOF(tagBuffer) - 1 &&
                   src[pos] != wxT('>') && !wxIsspace(src[pos]);
                   i++, pos++ )
             {
